+++ CHANGELOG.md
@@ -20,7 +20,12 @@ Not quite compatible with Octave yet. See #43 [octave support](https://github.co
 
 ### 0.9.38-dev
 
+* Fix issue where a Simulink model with self-contained links would not
+  parse the referenced systems. Only the top-level system would be
+  parsed and references were never followed.
 
+* Fix issue in `mh_trace` where sometimes LOBSTER tags could be
+  generated which contained spaces.
 
 ### 0.9.37